refactor(voice): split audio capture pipeline
